test(audit): give secret_in_environment and unpinned_image their own test files (#1716)
## Summary

- Splits `internal/audit/checks_more_tests.rs`, which had reached 432
code lines against a soft limit of 300, into three files by check.
- Pure move. No test renamed, merged, deleted or reworded.
- Fixes #1715.

## Changes

- New `internal/audit/checks_secret_env_tests.rs` and
`internal/audit/checks_unpinned_image_tests.rs`, registered in
`checks.rs` alongside the two module declarations already there.
- `checks_more_tests.rs` keeps the checks that did not move, with its
`//!` header corrected so it no longer describes contents that left.
- `checks_tests.rs` untouched.

The two checks that moved also had their tests scattered:
`secret_in_environment` sat in two separate blocks and `unpinned_image`
in two more, because the tests added with #1713 and #1714 were appended
at the end rather than beside their siblings. Each check now reads top
to bottom in one place.

## Test plan

- [x] `cargo fmt --all --check` is clean
- [x] `cargo clippy --locked --all-targets --all-features -- -D
warnings` is clean
- [x] `cargo test --locked --all-features --workspace` passes locally
- [x] `./tests/shell/*.test.sh` pass locally
- [x] shellcheck is clean on every tracked `*.sh`
- [ ] Podman lane ran (live Podman 5 and Podman 6, in the PR's own check
run)
- [ ] Asset-contract fixtures ran

**The control for a move is the test-name list, not the suite.** A
refactor that quietly drops a test goes green exactly like one that does
not. So the invariant checked here is that the set of bare test function
names is identical before and after, compared with:

```
cargo test --locked --all-features --bin podup -- --list \
  | grep '^audit::' | sed 's/.*:://' | sed 's/: test$//' | sort
```

Fully qualified names do change, since the module path moves; the bare
names must not.

This PR adds no check, so there is no control to delete and watch fail.
The move itself is verified by the list above rather than by a new
assertion.

- [x] New or changed checks were verified by deleting the control and
watching them fail

## Checklist

- [x] Targets `develop`
- [x] Commits are signed off (DCO, `git commit -s`)
- [x] Commits are signed
- [x] Labels applied
- [x] No secrets, keys or credentials in code, logs or fixtures
- [x] Docs updated if behaviour changed (no behaviour change)
- [x] `Cargo.toml` and `debian/changelog` are at the same version

## Related issues

Refs #1715.
